Linux registery bug fixed: Ubuntu Version: 20.04-LTS

+++ src/linux/index.js
@@ -62,6 +62,11 @@ const register = async (options, cb) => {
         const desktopFilePath = join(__dirname, '../../temp', desktopFileName);
         const desktopTemplate = join(__dirname, './templates', 'desktop.ejs');
         const scriptTemplate = join(__dirname, './templates', 'script.ejs');
+        const setDefaultTemplate = join(
+            __dirname,
+            './templates',
+            'default.ejs'
+        );
         const scriptFilePath = join(__dirname, '../../temp', 'script.sh');
 
         command = await preProcessCommands(protocol, command, scriptRequired);
@@ -90,6 +95,17 @@ const register = async (options, cb) => {
         });
         fs.writeFileSync(scriptFilePath, scriptContent);
 
+        const setDefaultCommand = await new Promise((resolve, reject) => {
+            ejs.renderFile(
+                setDefaultTemplate,
+                { protocol, desktopFileName, desktopFilePath },
+                function (err, str) {
+                    if (err) return reject(err);
+                    resolve(str);
+                }
+            );
+        });
+
         const chmod = await new Promise((resolve) =>
             shell.exec('chmod +x ' + scriptFilePath, (code, stdout, stderr) =>
                 resolve({ code, stdout, stderr })
@@ -108,6 +124,19 @@ const register = async (options, cb) => {
         );
         if (scriptResult.code != 0 || scriptResult.stderr)
             throw new Error(scriptResult.stderr);
+
+        const setDefaultResult = await new Promise((resolve) =>
+            shell.exec(
+                setDefaultCommand,
+                {
+                    silent: true
+                },
+                (code, stdout, stderr) => resolve({ code, stdout, stderr })
+            )
+        );
+
+        if (setDefaultResult.code != 0 || setDefaultResult.stderr)
+            throw new Error(setDefaultResult.stderr);
 
         fs.unlinkSync(scriptFilePath);
     } catch (e) {
